requirements.txt
fastapi
uvicorn
Pillow
piexif
opencv-contrib-python-headless
numpy
onnxruntime
python-multipart
